backend/movementControl.py
import serial  # If using serial communication
import logging

logging.basicConfig(level=logging.ERROR)

# Define serial port and baud rate (replace with your actual values)
SERIAL_PORT = "/dev/ttyAMA10"
BAUD_RATE = 115200

try:
    ser = serial.Serial(SERIAL_PORT, BAUD_RATE)
except serial.SerialException:
    logging.error("Serial communication not available.")
    ser = None

def move_forward():
    if ser:
        ser.write(b'f')  # Command for forward (replace if needed)

def move_backward():
    if ser:
        ser.write(b'b')  # Command for backward (replace if needed)

def move_left():
    if ser:
        ser.write(b'l')  # Command for left (replace if needed)

def move_right():
    if ser:
        ser.write(b'r')  # Command for right (replace if needed)

# ... other movement functions (turn_left, turn_right)